Adding Resolver#keys method
Given that the Resolver#key? method existed, I'm adding a mechanism to
inspect the container. I envision that this inspection is helpful from
a configuration validation and visualization stand-point. I am looking
to assert that my container has a handful of registrations so I can
begin writing further interface documentation for those dependencies.

Closes #12