Allow items to be initialized with a proc and then memoized
This commits adds a :memoize option to items. A memoized item
initialized with a proc (or block) will call the proc to get the value
only once, and will then reuse the item on further calls.

The main usage of this is allowing a number of dependencies to be
registered in the container without caring about their order (e.g. when
a dependency depends on another), but still reusing the same objects,
avoiding the need to instance them on each call to resolve, as is the
default when not using memoized.