Modernize rspec configuration
Generated a new spec_helper with rspec --init (after renaming the old
one) and merged with the existing spec_helper.

Modern rspec releases have a lot of useful options that are turned off
by default for backwards compatibility reasons, but that are rather
useful, so a newly-created configuration includes their
currently-recommended defaults.